+++ packages/commonwealth/client/scripts/views/pages/view_proposal/index.tsx
@@ -40,7 +40,7 @@ import {
 import { Prefetch } from './types';
 import { TipDetail } from '../tip_detail';
 import { ProposalBody } from './proposal_body';
-import { ProposalSidebar } from './proposal_sidebar';
+import { ThreadSidebar } from '../view_thread/thread_sidebar';
 
 class ViewProposalPage
   implements
@@ -511,7 +511,7 @@ class ViewProposalPage
               />
             )}
             {hasSidebar && this.tabSelected === 'viewSidebar' && (
-              <ProposalSidebar
+              <ThreadSidebar
                 isAdmin={isAdmin}
                 isAdminOrMod={isAdminOrMod}
                 isAuthor={isAuthor}
@@ -541,7 +541,7 @@ class ViewProposalPage
               viewCount={viewCount}
             />
             {hasSidebar && (
-              <ProposalSidebar
+              <ThreadSidebar
                 isAdmin={isAdmin}
                 isAdminOrMod={isAdminOrMod}
                 isAuthor={isAuthor}
