+++ packages/commonwealth/client/scripts/views/pages/view_proposal/proposal_body.tsx
@@ -30,7 +30,7 @@ import {
   ProposalBodyText,
   ProposalHeaderStage,
 } from './proposal_components';
-import { SharePopover } from './share_popover';
+import { SharePopover } from '../../components/share_popover';
 import { getThreadSubScriptionMenuItem } from '../discussions/discussion_row_menu';
 import { CWPopoverMenu } from '../../components/component_kit/cw_popover/cw_popover_menu';
 import { CWIconButton } from '../../components/component_kit/cw_icon_button';
