+++ packages/commonwealth/client/styles/pages/view_thread/linked_proposals_card.scss
