+++ packages/commonwealth/client/styles/pages/view_thread/linked_threads_card.scss
