+++ packages/commonwealth/client/styles/pages/view_thread/poll_editor_card.scss
